/ppbox/branches/1.2.0/src/ppbox/avformat/mp4 增加创建 mp4 文件功能，子 box 改成指针数组 emailto: box-dev
git-svn-id: http://svn01/repositories/r01/ppbox/branches/1.2.0@8020 b3a1389e-2159-5c43-9038-463bf877d0d9

+++ mp4/box/Mp4BoxHeader.h
@@ -88,7 +88,7 @@ namespace ppbox
                 type = id;
             }
 
-            boost::uint64_t head_size() const
+            boost::uint32_t head_size() const
             {
                 return (size == 1) ? 16 : 8;
             }
